+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/components/ButtonsFilter.java
@@ -83,6 +83,14 @@ final class ButtonsFilter extends Filter {
                 new ByteArrayFilterGroup(
                         Settings.HIDE_STOP_ADS_BUTTON,
                         "yt_outline_slash_circle_left"
+                ),
+                new ByteArrayFilterGroup(
+                        Settings.HIDE_HYPE_BUTTON,
+                        "yt_outline_star_shooting"
+                ),
+                new ByteArrayFilterGroup(
+                        Settings.HIDE_PROMOTE_BUTTON,
+                        "yt_outline_megaphone"
                 )
         );
     }